feat: rename 'manage soft liquidation' to 'manage liquidation'
