+++ apps/main/src/lend/components/PageLendMarket/LendMarketPage.tsx
@@ -9,7 +9,7 @@ import { type MarketUrlParams } from '@/lend/types/lend.types'
 import { getCollateralListPathname, parseMarketParams } from '@/lend/utils/utilsRouter'
 import { MarketContextProvider } from '@/llamalend/features/market-context'
 import { PositionDetailsComposite } from '@/llamalend/features/market-position-details'
-import { useIsInSoftLiquidation } from '@/llamalend/features/market-position-details/hooks/useUserLiquidationStatus'
+import { useIsInLiquidation } from '@/llamalend/features/market-position-details/hooks/useUserLiquidationStatus'
 import { useUserCollateralEvents } from '@/llamalend/features/user-position-history/hooks/useUserCollateralEvents'
 import { useLlamaMarket } from '@/llamalend/hooks/useLlamaMarket'
 import { getControllerAddress, getTokens, hasResetPosition } from '@/llamalend/llama.utils'
@@ -66,10 +66,7 @@ export const LendMarketPage = () => {
     network,
   })
   const showReset = useMarketResetPosition() && hasResetPosition(market)
-  const { data: isSoftLiquidation, isLoading: isSoftLiquidationLoading } = useIsInSoftLiquidation(
-    queryParams,
-    !!loanExists,
-  )
+  const { data: isLiquidation, isLoading: isLiquidationLoading } = useIsInLiquidation(queryParams, !!loanExists)
 
   const error = marketError ?? apiMarket.error
   return error ? (
@@ -87,13 +84,13 @@ export const LendMarketPage = () => {
           content:
             !isLoading &&
             !isLoanExistsLoading &&
-            !isSoftLiquidationLoading &&
+            !isLiquidationLoading &&
             (loanExists ? (
               <ManageLoanTabs
                 onPricesUpdated={setPreviewPrices}
                 collateralEvents={collateralEvents}
                 showReset={showReset}
-                isSoftLiquidation={!!isSoftLiquidation}
+                isLiquidation={!!isLiquidation}
               />
             ) : (
               <CreateLoanTabs onPricesUpdated={setPreviewPrices} />
