+++ apps/main/src/llamalend/features/manage-loan/components/RepayForm.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react'
 import { LEVERAGE } from '@/llamalend/constants'
 import { RepayLoanInfoList } from '@/llamalend/features/borrow/components/RepayLoanInfoList'
+import { AlertRepayDebtToIncreaseHealth } from '@/llamalend/features/manage-liquidation/ui/alerts/AlertRepayDebtToIncreaseHealth'
 import { RepayTokenList, type RepayTokenListProps } from '@/llamalend/features/manage-loan/components/RepayTokenList'
 import { RepayTokenOption, useRepayTokens } from '@/llamalend/features/manage-loan/hooks/useRepayTokens'
-import { AlertRepayDebtToIncreaseHealth } from '@/llamalend/features/manage-soft-liquidation/ui/alerts/AlertRepayDebtToIncreaseHealth'
 import type { UserCollateralEvents } from '@/llamalend/features/user-position-history/hooks/useUserCollateralEvents'
 import { hasLeverageValue } from '@/llamalend/llama.utils'
 import type { NetworkDict } from '@/llamalend/llamalend.types'
@@ -26,7 +26,7 @@ import { CRVUSD } from '@ui-kit/utils'
 import { Form } from '@ui-kit/widgets/DetailPageLayout/Form'
 import { FormAlerts, HighPriceImpactAlert } from '@ui-kit/widgets/DetailPageLayout/FormAlerts'
 import { shouldBlockTransaction } from '@ui-kit/widgets/DetailPageLayout/price-impact.util'
-import { useCrvSwapUrl } from '../../manage-soft-liquidation/hooks/useCrvSwapUrl'
+import { useCrvSwapUrl } from '../../manage-liquidation/hooks/useCrvSwapUrl'
 import { useMarketContext } from '../../market-context'
 import { useRepayForm } from '../hooks/useRepayForm'
 import { useTokenAmountConversion } from '../hooks/useTokenAmountConversion'
