+++ apps/main/src/llamalend/features/market-position-details/hooks/useUserLiquidationStatus.ts
@@ -24,7 +24,7 @@ export const useLiquidationStatus = (params: UserMarketParams, enabled?: boolean
       ),
   )
 
-export const useIsInSoftLiquidation = (params: UserMarketParams, enabled?: boolean) =>
+export const useIsInLiquidation = (params: UserMarketParams, enabled?: boolean) =>
   mapQuery(useLiquidationStatus(params, enabled), status =>
     ['softLiquidation', 'hardLiquidation', 'fullyConverted'].includes(status),
   )
